git-svn-id: https://mattari.mydns.jp/repos/android_apps/trunk/NitechBBSrv@639 ebda3f91-3b32-42d3-bb0a-b50f2d3b12aa

+++ app/models/service/model/BBUserClusterModelService.java
@@ -66,7 +66,7 @@ public class BBUserClusterModelService implements ModelService<Long, BBUserClust
 							.findUnique();
 	}
 	
-	public Set<BBUserCluster> findSetByClusterDepth(long clusterDepth) {
+	public Set<BBUserCluster> findSetByClusterDepth(int clusterDepth) {
 		return BBUserCluster.find
 							.where()
 								.eq("clusterDepth", clusterDepth)
